Rename redux to modules
