+++ frontend/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render } from 'react-dom'
-import createStore from 'redux'
+import createStore from 'modules'
 import Root from 'components/Root'
 
 const store = createStore()
