+++ frontend/src/modules/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux'
-import configureStore from 'redux/createStore'
+import configureStore from 'modules/createStore'
 
 export default () => {
   const rootReducer = combineReducers({
