+++ frontend/src/modules/noteRedux.js
